+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 1.14.0-rc.1
+
+Released 2025-Nov-13
+
+* Updated OpenTelemetry core component version(s) to `1.14.0`.
+  ([#3403](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3403))
+
 ## 1.13.0-rc.1
 
 Released 2025-Oct-28