+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -2,8 +2,18 @@
 
 ## Unreleased
 
+## 1.14.0-beta.1
+
+Released 2025-Nov-13
+
 * Add support for .NET 10.0.
   ([#2822](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2822))
+
+* Update .NET 10.0 NuGet package versions from `10.0.0-rc.2.25502.107` to `10.0.0`.
+  ([#3403](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3403))
+
+* Updated OpenTelemetry core component version(s) to `1.14.0`.
+  ([#3403](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3403))
 
 ## 1.13.0-beta.1
 
@@ -30,7 +40,7 @@ Released 2025-Oct-22
   * The default behavior (in the absence of one of these values) is to continue
   emitting the same database semantic conventions that were emitted in
   the previous version.
-  * Note: this option will be be removed after the new database
+  * Note: this option will be removed after the new database
   semantic conventions are marked stable. At which time this
   instrumentation can receive a stable release, and the old database
   semantic conventions will no longer be supported. Refer to the